Added keyword "desctructive" which will remove the options from the array being parsed when they are found.
An example use case of when this would be appropriate would be a
program with the following usage:
usage: program [-abc] required_file

Now with the destructive option set the user can expect all the
right options in Choice.choices and expect his ARGV to contain
only the filename.  This can probably be improved.

Joseph Pecoraro